Allow trailing slashes in URIs by adding the NormalizePathLayer
Problem
-------
Some clients may append a trailing slash to URIs, either intentionally
or accidentally. We should support both routes that do and don't have
trailing slashes for maximum compatibility.

Solution
--------
Normally, when we define a route, it will not have a trailing slash. We
can add support for trailing slashes by adding the `NormalizePathLayer`,
which will automatically remove the trailing slash from the requested
URI, allowing it to map to the route we created without a trailing
slash.