Fix the doc test for `src/controller/mod.rs`
Problem
-------
The same rust code in `src/controller/mod.rs` does not compile with a
fresh clone of `loco` because it is using some methods and imports that
are only available when the `channels` feature is enabled.

Solution
--------
Update the sample code to guard the offending blocks of code with a
`cfg` guard.